+++ src/devices/MotorHat/MotorPinProvider.cs
@@ -1,4 +1,7 @@
-﻿using System;
+﻿// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+
+using System;
 
 namespace Iot.Device.MotorHat
 {
@@ -16,7 +19,7 @@ namespace Iot.Device.MotorHat
     }
 
     /// <summary>
-    /// Abstract base class for creating DCMotor instances
+    /// Static class providing known <see cref="IMotorPinProvider"/> instances
     /// </summary>
     public static class MotorPinProvider
     {
@@ -39,55 +42,5 @@ namespace Iot.Device.MotorHat
         /// Default <see cref="IMotorPinProvider"/>
         /// </summary>
         public static readonly IMotorPinProvider Default = AdaFruit;
-    }
-
-    /// <summary>
-    /// <see cref="MotorPinProvider"/> implementation for AdaFruit/Aliexpress
-    /// </summary>
-    /// <remarks>
-    /// These correspond to motor hat screw terminals M1, M2, M3 and M4.
-    /// </remarks>
-    public class AdafruitMotorPinProvider : IMotorPinProvider
-    {
-        /// <summary>
-        /// Gets the <see cref="MotorPins"/> for motor at the specified <paramref name="index"/>
-        /// </summary>
-        /// <param name="index">The index of the motor to get pins for</param>
-        /// <returns></returns>
-        public MotorPins GetPinsForMotor(int index)
-        {
-            return index switch
-            {
-                1 => new MotorPins(8, 9, 10),
-                2 => new MotorPins(13, 12, 11),
-                3 => new MotorPins(2, 3, 4),
-                4 => new MotorPins(7, 6, 5),
-                _ => throw new ArgumentException(nameof(index), $"MotorHat Motor must be between 1 and 4 inclusive. {nameof(index)}: {index}")
-            };
-        }
-    }
-
-    /// <summary>
-    /// <see cref="MotorPinProvider"/> implementation for AdaFruit/Aliexpress
-    /// </summary>
-    /// <remarks>
-    /// These correspond to motor hat screw terminals M1 and M2
-    /// </remarks>
-    public class WaveshareMotorPinProvider : IMotorPinProvider
-    {
-        /// <summary>
-        /// Gets the <see cref="MotorPins"/> for motor at the specified <paramref name="index"/>
-        /// </summary>
-        /// <param name="index">The index of the motor to get pins for</param>
-        /// <returns></returns>
-        public MotorPins GetPinsForMotor(int index)
-        {
-            return index switch
-            {
-                1 => new MotorPins(0, 1, 2),
-                2 => new MotorPins(5, 4, 3),
-                _ => throw new ArgumentException(nameof(index), $"MotorHat Motor must be either 1 or 2. {nameof(index)}: {index}")
-            };
-        }
     }
 }
